+++ datasources/dataflux/dataflux.go
@@ -5,12 +5,14 @@ import (
     "github.com/influxdata/influxdb/client/v2"
 )
 
+// DataFlux - datasource keeping IOT data
 type DataFlux struct {
     Client          client.Client
     BatchPoints     client.BatchPoints
     isConnected     bool
 }
 
+// New - creates new entity of DataFlux
 func New(dbURI string) DataFlux {
     isConnected := false
 
@@ -44,15 +46,18 @@ func New(dbURI string) DataFlux {
     }
 }
 
+// IsConnected - checks is DataFlux instance connected
 func (df DataFlux) IsConnected() bool {
     return df.isConnected
 }
 
+// GetData - gets data from DataFlux instance based on passed query
 func (df DataFlux) GetData(q client.Query) (*client.Response, error) {
     res, err := df.Client.Query(q)
     return res, err
 }
 
+// AddData - adds data to instance of DataFlux
 func (df DataFlux) AddData(pt *client.Point) error {
     df.BatchPoints.AddPoint(pt)
     err := df.Client.Write(df.BatchPoints)
