+++ datasources/state/state.go
@@ -8,13 +8,14 @@ import (
     "github.com/smart-evolution/smarthome/models/agent"
 )
 
-// State
+// State - data source which keeps short data in memory
 type State struct {
     isAlerts    bool
     sendAlert   bool
     agents      []agent.Agent
 }
 
+// New - creates new instance of State
 func New() State {
     return State{
         isAlerts: false,
@@ -23,22 +24,27 @@ func New() State {
     }
 }
 
+// Agents - returns list of available agents
 func (s State) Agents() []agent.Agent {
     return s.agents
 }
 
+// SetIsAlerts - setted for `isAlerts`
 func (s State) SetIsAlerts(i bool) {
     s.isAlerts = i
 }
 
+// IsAlerts - getter for `isAlerts`
 func (s State) IsAlerts() bool {
     return s.isAlerts
 }
 
+// SetSendAlert - setter for `sendAlert`
 func (s State) SetSendAlert(i bool) {
     s.sendAlert = i
 }
 
+// SendAlert - getter for `sendAlert`
 func (s State) SendAlert() bool {
     return s.sendAlert
 }
@@ -51,6 +57,7 @@ func (s State) addAgent(id string, name string, url string, agentType string) {
     s.agents = append(s.agents, agent)
 }
 
+// SetupAgents - setups agents based on hardware config file
 func (s State) SetupAgents(hardwareFile string) {
     agentsCnf, err := ioutil.ReadFile(hardwareFile)
